fix(packages/sui-ssr): move fn to utils
