+++ packages/sui-ssr/server/utils/factory.js
@@ -112,14 +112,38 @@ export default ({path, fs, config: ssrConf = {}}) => {
     return url
   }
 
+  const getStyleHrefBy = ({pageName}) => {
+    const assetsFile = ssrConf.assetsManifest
+
+    return assetsFile && assetsFile[`${pageName}.css`]
+  }
+
+  const createStylesFor = ({pageName, async = false}) => {
+    const appStyles = ssrConf.createStylesFor.appStyles
+    const shouldCreatePageStyles = ssrConf.createStylesFor.createPagesStyles
+
+    const stylesheets = [
+      appStyles && getStyleHrefBy({pageName: appStyles}),
+      shouldCreatePageStyles && getStyleHrefBy({pageName})
+    ].filter(Boolean)
+
+    const attributes = async ? ssrConf.ASYNC_CSS_ATTRS : ''
+    const stylesHTML = stylesheets
+      .map(style => `<link rel="stylesheet" href="${style}" ${attributes}>`)
+      .join('')
+
+    return stylesHTML
+  }
+
   return {
-    isMultiSite,
-    hostFromReq,
-    siteByHost,
-    useStaticsByHost,
-    readHtmlTemplate,
     buildRequestUrl,
+    createStylesFor,
+    hostFromReq,
+    hrTimeToMs,
+    isMultiSite,
     publicFolder,
-    hrTimeToMs
+    readHtmlTemplate,
+    siteByHost,
+    useStaticsByHost
   }
 }
